requirements.txt
GitPython==3.1.24
Jinja2==3.0.2
google-cloud-translate==3.5.0
pypandoc==1.6.4